chore(*): Moved directives to the shorter namespace `ph`
placeholderImage is now phImg and placeholderText is now phTxt. I also
added a missing section to the phImg demo page that goes with the demo
javascript provided, and added an additional section to show how the
directive's flexibility is enhanced through the use of CSS classes, such
as Twitter Bootstrap's `img-rounded`, `img-polaroid`, and `img-circle`.